Create klc.pl
Hello. Yes, I originally authored this script. I do appreciated the improvements which Dutch contributed.
I have no idea if this is the latest revision floating around the community or not -- and I might need to dig deep for the pre-dutch changes.
Resurrected from http://www.netstumbler.org/scripts/kismack-kismet-merge-and-convert-to-kml-t18959-30.html before it is lost forever